requirements.txt
requests
pandas
unidecode
mysql-connector-python
python-dotenv